+++ core/amber/src/main/python/core/models/payload.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from pyarrow.lib import Table
 from typing import List
 
-from core.models.schema import Schema
+from core.models.schema.schema import Schema
 from core.models.tuple import Tuple
